improvement(heroko): utilize seperate config on migrate as needed
